Bugfixes for compatibility with MATLAB 2014 autotrace: makeplots window had full file path in title.

+++ Traces/loadTraces.m
@@ -183,7 +183,7 @@ type = dataTypes{dataType};
 
 % Read metadata contents
 if strcmp(type,'cell'),
-    metadata = cell(szMetadata);
+    metadata = cell( double(szMetadata) );
     for i=1:nMetadata,
         metadata{i} = readMetadata(fid);
     end
